fix: Refactor the load method missing parts
